Import function for the automated chamber ECOFlux (GAIA2TECH) synced with multiple chambers and instruments

Imports single raw gas measurement files from the automated chamber ECOFlux (GAIA2TECH) linked with up to two instruments and up to five gases (by default used with LI-7810: CO2, CH4 and H2O / LI7820: N2O and H2O).

GAIA_import(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  background = FALSE,
  save = FALSE,
  prec = c(3.5, 0.6, 0.4, 45, 45),
  CH.col = "COM5A0",
  SWC.col = "1C08_Soil.Moisture",
  Tsoil.col = "1C07_Soil.Temperature",
  Tcham.col = "2C07_Chamber.Temperature",
  PAR.col = "3C07_Sunlight",
  Op.stat.col = "0C06_OperatingStatus",
  CO2.col = "XT2C05_CO2",
  CH4.col = "XT2C04_CH4",
  H2O1.col = "XT2C06_H2O",
  H2O1.name = "H2O_LI7810",
  N2O.col = "XT3C04_N2O",
  H2O2.col = "XT3C05_H2O",
  H2O2.name = "H2O_LI7820",
  sep = "\t",
  skip = 1
)

import.GAIA(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  background = FALSE,
  save = FALSE,
  prec = c(3.5, 0.6, 0.4, 45, 45),
  CH.col = "COM5A0",
  SWC.col = "1C08_Soil.Moisture",
  Tsoil.col = "1C07_Soil.Temperature",
  Tcham.col = "2C07_Chamber.Temperature",
  PAR.col = "3C07_Sunlight",
  Op.stat.col = "0C06_OperatingStatus",
  CO2.col = "XT2C05_CO2",
  CH4.col = "XT2C04_CH4",
  H2O1.col = "XT2C06_H2O",
  H2O1.name = "H2O_LI7810",
  N2O.col = "XT3C04_N2O",
  H2O2.col = "XT3C05_H2O",
  H2O2.name = "H2O_LI7820",
  sep = "\t",
  skip = 1
)

Arguments

inputfile

character string; the name of a file with the extension .csv

date.format

character string; specifies the date format found in the raw data file. Choose one of the following: "dmy", "ymd", or "mdy". Default is "ymd", as it is the date format from the example data file provided.

timezone

character string; a time zone in which to import the data to POSIXct format. Default is "UTC". Note about time zone: it is recommended to use the time zone "UTC" to avoid any issue related to summer time and winter time changes.

pivot

character string; either "long" or "wide". If pivot = "long", each column containing information about Tsoil, Tcham, SWC, PAR and operating status (Op.stat) will be saved in a single column per parameter. If pivot = "wide", the default display of one column per chamber per parameter will be used.

active

logical; if active = TRUE, preserve data for active chambers only.

flag

numeric vector; indicates the operating status that should be used for the flux calculation. Default is flag = c(7,11), where 7 indicates "Chamber Idle Closed Clear" and 11 indicates "Chamber Idle Closed Dark".

background

logical; if background = FALSE, removes all data from activ.cham == "Background".

save

logical; if save = TRUE, saves the file as an .RData file in a RData folder in the current working directory. If save = FALSE, returns the file in the Console, or load in the Environment if assigned to an object.

prec

numerical vector; the precision of the instrument for each gas, in the following order: "CO2dry_ppm", "CH4dry_ppb", "N2Odry_ppb" and "H2O_ppm" from each analyzer. The default is prec = c(3.5, 0.6, 0.4, 45, 45). For each gas missing, the value is ignored.

Op.stat.col, PAR.col, Tcham.col, Tsoil.col, SWC.col, CH.col

character string; a pattern to match all columns that fit the corresponding parameter. For example, all columns containing the pattern "3C07_Sunlight" will be renamed with the pattern "_PAR". Then, if pivot = "long", all columns with the pattern "_PAR" will be merged together.

CO2.col, CH4.col, H2O1.col, N2O.col, H2O2.col

character string; a pattern to match the columns containing the corresponding gas measurements. H2O1.col must be the same instrument as CO2.col and CH4.col, and H2O2.col must be the same instrument as N2O.col. If columns are absent from the raw data, the arguments are ignored.

H2O1.name, H2O2.name

character string; how to name the water vapor columns for each instrument. Ideally, use H201.name = "H2O_instrument.name". For example, H201.name = "H2O_LI7810".

sep

character string defining the field separator character. Values on each line of the file are separated by this character. By default, sep = "\t" for tabulation.

skip

integer; the number of lines of the data file to skip before beginning to read data. By default, skip = 1.

Value

A data frame containing raw data from the automated chamber ECOFlux (GAIA2TECH)

Details

This function has been designed for the automated chamber ECOFlux (GAIA2TECH) synced with up to 9 chambers and two GHG analyzers from LI-COR (LI-7810: CO2, CH4 and H2O / LI7820: N2O and H2O). If this function could be useful for you, but does not meet your needs, please contact the maintainer of this package for potential adaptations.

In date.format, the date format refers to a date found in the raw data file, not the date format in the file name. For the instrument GAIA the date is found in the column "Titles:".

The arguments PAR.col, Tcham.col, Tsoil.col and SWC.col correspond to different types of probes linked to the ECOFlux chamber: PAR, chamber temperature, soil temperature and soil water content volumetric, respectively. The argument Op.stat.col corresponds to the columns Operating Status of each chamber. The argument CH.col indicates a character string preceding the chamber number for each column of the raw data. For example, the column "COM5A010C06_OperatingStatus" in the raw data file contains the Operating Status for the chamber 1 if CH.col = "COM5A0" and Op.stat.col = "0C06_OperatingStatus". If columns are absent from the raw data, the arguments are ignored.

Note that this function was designed for the following units in the raw file:

ppm for CO2 and H2O

ppb for CH4 and N2O

kPa for pressure

volumetric water content (%) for soil moisture

Celsius for temperature

µmol photons m-2s-1 for PAR

If your instruments use different units, either convert the units after import, change the settings on your instrument, or contact the maintainer of this package for support.

As opposed to the other import functions, there is no option to "keep_all" with this instrument. If you would like to import additional data using this function, please contact the maintainer of this package for support.

The precision of the instrument is needed to restrict kappa-max (k.max) in the non-linear flux calculation (HM.flux). Kappa-max is inversely proportional to instrument precision. If the precision of your instrument is unknown, it is better to use a low value (e.g. 1 ppm for CO2 and H2O, or 1 ppb for CH4 and N2O) to allow for more curvature, especially for water vapor fluxes, or very long measurements, that are normally curved. The default values given for instrument precision are the ones found online for the latest models of the LI-7810 and LI-7820, available at the time of the creation of this function (11-2023). Five values are required for this argument and their position in the vector is important. If you are using the function with less than five gases, the other values are simply ignored.

See also

Examples

# Load file from downloaded package
file.path <- system.file("extdata", "GAIA/GAIA.csv", package = "goFlux")

# Run function
imp.GAIA <- import.GAIA(inputfile = file.path)