Non-linear model for flux calculation

Estimates a gas flux with a non-linear regression, the Hutchinson and Mosier model, and then extracts values from the model fit.

HM.flux(
  gas.meas,
  time.meas,
  flux.term,
  k.max,
  k.min,
  Ct = NULL,
  C0 = NULL,
  k.mult = 1
)

Arguments

gas.meas

numerical vector containing gas measurements (ppm or ppb)

time.meas

numerical vector containing time stamps (seconds)

flux.term

numerical value; flux term calculated with the function flux.term

k.max

numerical value; kappa-max is the maximal curvature allowed in the Hutchinson and Mosier model. Calculated with the k.max function.

k.min

numerical value; a lower boundary value for kappa in the HM model. To allow for a log-like curvature, set k.min below 0 (ex. -1), otherwise it should be just above 0 (ex. 10^-8). k.min cannot be 0 as this would result in a singular gradient.

Ct

numerical value; gas concentration after t seconds (ppm or ppb).

C0

numerical value; initial gas concentration at time 0 second (ppm or ppb).

k.mult

numerical value; a multiplier for the allowed k.max. Default is no multiplier (k.mult = 1).

Value

Returns a data frame with 11 columns: non-linear flux estimate, initial gas concentration (HM.C0), the assumed concentration of constant gas source below the surface (HM.Ct), slope at t=0 (HM.slope), mean absolute error (HM.MAE), root mean square error (HM.RMSE), Akaike's information criterion corrected for small sample size (LM.AICc), standard error (HM.SE), relative standard error (HM.se.rel), coefficiend of determination (HM.r2), and curvature (kappa; (HM.k)).

Details

Flux estimate units are µmol m-2s-1 (if initial concentration is ppm, e.g. CO2dry_ppm) and nmol m-2s-1 (if initial concentration is ppb, e.g. CH4dry_ppb).

The Hutchinson and Mosier model (HM) is a non-linear model, whose curvature is controlled by the parameter kappa. A large kappa returns a strong curvature. A maximum threshold for this parameter, kappa-max (k.max), can be calculated from the linear flux estimate (LM.flux), the minimal detectable flux (MDF) and the time of chamber closure. This limit of kappa-max is included in the goFlux function, so that the non-linear flux estimate cannot exceed this maximum curvature.

The HM slope is multiplied by a flux.term which is used to correct for water vapor inside the chamber, as well as convert the units of the slope to obtain a flux estimate in nmol or µmol m-2s-1.

In Ct and C0, typical values for soil respiration and 180 seconds of chamber closure are as follows for each gastype:

CO2dry_ppm: Ct=1000 ; C0=380

CH4dry_ppb: Ct=1800 ; C0=2200

N2Odry_ppb: Ct=500 ; C0=250

H2O_ppm: Ct=20000 ; C0=10000

Values of Ct and C0 will vary depending on ecosystem type and chamber application scheme.

References

Hüppi et al. (2018). Restricting the nonlinearity parameter in soil greenhouse gas flux calculation for more reliable flux estimates. PloS one, 13(7), e0200876.

Hutchinson and Mosier (1981). Improved soil cover method for field measurement of nitrous oxide fluxes. Soil Science Society of America Journal, 45(2), 311-316.

See also