Import function for LI-COR GHG analyzer LI-7810

Imports single raw gas measurement files from the LI-COR 7810 (CO2, CH4 and H2O GHG analyzer)

LI7810_import(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  save = FALSE,
  keep_all = FALSE,
  prec = c(3.5, 0.6, 45)
)

import.LI7810(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  save = FALSE,
  keep_all = FALSE,
  prec = c(3.5, 0.6, 45)
)

Arguments

inputfile

character string; the name of a file with the extension .data or .txt

date.format

character string; specifies the date format found in the raw data file. Choose one of the following: "dmy", "ymd", or "mdy". Default is "ymd", as it is the date format from the example data file provided.

timezone

character string; a time zone in which to import the data to POSIXct format. Default is "UTC". Note about time zone: it is recommended to use the time zone "UTC" to avoid any issue related to summer time and winter time changes.

save

logical; if save = TRUE, saves the file as an .RData file in a RData folder in the current working directory. If save = FALSE, returns the file in the Console, or load in the Environment if assigned to an object.

keep_all

logical; if keep_all = TRUE, keep all columns from the raw file. The default is keep_all = FALSE, and columns that are not necessary for gas flux calculation are removed.

prec

numerical vector; the precision of the instrument for each gas, in the following order: "CO2dry_ppm", "CH4dry_ppb" and "H2O_ppm". The default is prec = c(3.5, 0.6, 45).

Value

A data frame containing raw data from LI-COR GHG analyzer LI-7810.

Details

In date.format, the date format refers to a date found in the raw data file, not the date format in the file name. For the instrument LI-7810, the date is found in the column "DATE".

Note that this function was designed for the following units in the raw file:

ppm for CO2 and H2O

ppb for CH4

kPa for pressure

Celsius for temperature

If your LI-COR LI-7810 uses different units, either convert the units after import, change the settings on your instrument, or contact the maintainer of this package for support.

The precision of the instrument is needed to restrict kappa-max (k.max) in the non-linear flux calculation (HM.flux). Kappa-max is inversely proportional to instrument precision. If the precision of your instrument is unknown, it is better to use a low value (e.g. 1 ppm for CO2 and H2O, or 1 ppb for CH4) to allow for more curvature, especially for water vapor fluxes, or very long measurements, that are normally curved. The default values given for instrument precision are the ones found online for the latest model of this instrument available at the time of the creation of this function (11-2023).

See also

Examples

# Load file from downloaded package
file.path <- system.file("extdata", "LI7810/LI7810.data", package = "goFlux")

# Run function
imp.LI7810 <- import.LI7810(inputfile = file.path)