Import function for LI-COR smart chamber (LI-8200)

Imports single raw data files from the LI-COR smart chamber (LI-8200)

LI8200_import(inputfile, date.format = "ymd", timezone = "UTC", save = FALSE)

import.LI8200(inputfile, date.format = "ymd", timezone = "UTC", save = FALSE)

Arguments

inputfile

character string; the name of a file with the extension .json

date.format

character string; specifies the date format found in the raw data file. Choose one of the following: "dmy", "ymd", or "mdy". Default is "ymd", as it is the date format from the example data file provided.

timezone

character string; a time zone in which to import the data to POSIXct format. Default is "UTC". Note about time zone: it is recommended to use the time zone "UTC" to avoid any issue related to summer time and winter time changes.

save

logical; if save = TRUE, saves the file as an .RData file in a RData folder in the current working directory. If save = FALSE, returns the file in the Console, or load in the Environment if assigned to an object.

Value

A data frame containing raw data from the LI-COR Smart Chamber LI-8200.

Details

In date.format, the date format refers to a date found in the raw data file, not the date format in the file name. For the instrument LI-8200, the date is found under one of the measurements, next to '"Date":'.

Note that this function was designed for the following units in the raw file:

seconds for DeadBand

cm for Offset (collar height)

(cm2) for Area

(cm3) for volumes

mmol/mol for H2O

kPa for pressure

Celsius for temperature

If your Smart Chamber uses different units, either convert the units after import, change the settings on your instrument, or contact the maintainer of this package for support.

Important note about the total volume: it is calculated from the addition of the Area x Offset and the Chamber Volume. If a chamber volume is specified ("ChamVolume" in the raw file), that value is used, otherwise, the default value is 4244.1(cm3). For more precise results, remember to add the volume inside your GHG analyzer (IRGA volume; e.g., 6.4 (cm3) for the LI-7810 and the LI-7820), plus the volume inside the tubing connecting your instruments. For more details about volume estimation, download the example auxiliary file from the goFlux webpage.

In addition, soil temperature and moisture are read with a HydraProbe connected to the Smart Chamber. If you are using a different setup, please contact the maintainer of this package for support.

As opposed to the other import functions, there is no option to "keep_all" with the Smart Chamber. If you would like to import additional data using this function, please contact the maintainer of this package for support.

See also

Examples

# Load file from downloaded package
file.path <- system.file("extdata", "LI8200/LI8200.json", package = "goFlux")

# Run function
imp.LI8200 <- import.LI8200(inputfile = file.path)