Linear model for flux calculation

Estimates a gas flux with a linear regression, and then extracts values from the linear fit.

LM.flux(gas.meas, time.meas, flux.term)

Arguments

gas.meas

numerical vector containing gas measurements (ppm or ppb)

time.meas

numerical vector containing time stamps (seconds)

flux.term

numerical value; flux term calculated with the function flux.term

Value

Returns a data frame with 11 columns: linear flux estimate, initial gas concentration (LM.C0), final gas concentration (LM.Ct), slope of linear regression (LM.slope), mean absolute error (LM.MAE), root mean square error (LM.RMSE), Akaike information criterion corrected for small sample size (LM.AICc), standard error (LM.SE), relative standard error (LM.se.rel), r2 (LM.r2), and p-value (LM.p.val).

Details

Flux estimate units are µmol/m2s (if initial concentration is ppm, e.g. CO2dry_ppm) and nmol/m2s (if initial concentration is ppb, e.g. CH4dry_ppb).

See also