Minimal Detectable Flux (MDF)

The minimal detectable flux based on instrument precision and the measurement time.

MDF(p, t, flux.term)

Arguments

p

numerical; precision of the instrument (same units as measured gas; ex. ppm)

t

numerical; measurement time (enclosure time; seconds)

flux.term

numerical; flux term calculated with the function flux.term

Value

a numerical value

References

Christiansen et al. (2015). Comparison of CO2, CH4 and N2O soil-atmosphere exchange measured in static chambers with cavity ring-down spectroscopy and gas chromatography. Agricultural and Forest Meteorology, 211, 48-57.

See also