Automatic selection of deadband per measurement

Loops through multiple gas measurements and automatically define a suitable deadband per measurement.

auto.deadband(
  dataframe,
  gastype,
  db.min = 0,
  db.max = 60,
  obs.win = 90,
  mov.win = 10,
  min.r2 = 0
)

Arguments

dataframe

data.frame; output from import or ID functions.

gastype

character string; specifies which gas column should be used to detect inflection point. Must be one of the following: "CO2dry_ppm", "CH4dry_ppb", "COdry_ppb", "N2Odry_ppb", "NH3dry_ppb", "NOdry_ppb", "NO2dry_ppb" or "H2O_ppm".

db.min

numerical value (seconds); minimal deadband per measurement.

db.max

numerical value (seconds); maximal deadband per measurement.

obs.win

numerical value (seconds); maximal observation window to include in the linear models. If possible, keept at least 30 seconds between db.max and obs.win for more accurate results.

mov.win

numerical value (seconds); time increment between obs.win, from chamber closure to db.max.

min.r2

numerical value; minimum increase in R2 required to accept dropping more data points.

Value

A data.frame of the automatically selected deadband per UniqueID.

Details

For each measurement, a sequence of observation window is defined from chamber closure (+ mov.win) to obs.win. For each observation window, a linear model is calculated, from which the R2 is extracted. The 'best' deadband is then selected at the highest R2.

In db.min and db.max, negative values are accepted. This is especially applicable in cases where start.time already includes a deadband after chamber closure, and cham.close is not provided.

Note that, obs.win should represent the most linear part of the measurement, at the start of the measurement. For long chamber enclosures (more than a few minutes) select a longer obs.win more representative of your measurements. The default value of obs.win = 90s is based on previous observations that the most linear part of a measurement is often found within a short period (<2 minutes) at the start of the measurement, for small chamber sizes (~4L) (Johannesson et al., 2024). Similarly, the argument mov.win could be increased for very long chamber enclosures. Using a short mov.win (e.g., 1s) increases computational time, and is not recommended for very large data sets. The default value of db.max = 60 is based on dark chamber measurements on, e.g., soils. For Reco, a longer db.max (e.g., 90 seconds) is recommended to take into account a recovery phase of residual photosynthesis at the start of the measurement.

min.r2 can be used to restrict longer deadbands (or minimize dropping data points). For example, if an R2 of 0.990 is found with a deadband of 30 seconds, and an R2 of 0.992 with a deadband of 40 seconds (considering mov.win = 10), the best deadband is defined at 30 seconds if min.r2 = 0.005. By default, the highest R2 selects the best deadband (min.r2 = 0), but this may result in unnecessary loss of data points.

In gastype, the gas species listed are the ones for which this package has been adapted. Please write to the maintainer of this package for adaptation of additional gases.

References

Johannesson, C. F., Nordén, J., Lange, H., Silvennoinen, H., & Larsen, K. S. (2024). Optimizing the closure period for improved accuracy of chamber-based greenhouse gas flux estimates. Agricultural and Forest Meteorology, 359, 110289.

See also

Examples

data("manID.UGGA")
auto_db <- auto.deadband(manID.UGGA, "CO2dry_ppm")
crop_UGGA <- crop.meas(manID.UGGA, auto_db, deadband = "aux")