Automatically define measurements' window based on an auxiliary file

Uses an auxiliary file to identify start and end time of measurements, add a UniqueID per measurement, and merge other auxiliary data required with the goFlux function.

autoID(
  inputfile,
  auxfile = NULL,
  obs.length = NULL,
  deadband = 0,
  crop.end = 0,
  shoulder = 60,
  drop = TRUE
)

Arguments

inputfile

data.frame; output from import or align functions.

auxfile

data.frame; auxiliary data frame containing the columns start.time and UniqueID. start.time must contain a date and be in POSIXct format. The time zone must be the same as the POSIX.time in inputfile. The default time zone for the import functions is "UTC". A data frame from the Smart Chamber (LI-8200) can be used as an auxiliary file. In that case, chamID will be used instead of UniqueID, if UniqueID cannot be found.

obs.length

numerical; chamber closure time (seconds). Default is NULL. If obs.length is not provided, a column obs.length should be contained in auxfile or inputfile. Alternatively, obs.length will be calculated from start.time and cham.open or end.time if found in auxfile or inputfile.

deadband

numerical; time (seconds) at the start of the measurement, after chamber closure, to be removed from the time stamps used to estimate the flux.

crop.end

numerical; time (seconds) at the end of the measurement, before chamber opening, to be removed from the time stamps used to estimate the flux.

shoulder

numerical; time before and after measurement in observation window (seconds). Default is 60 seconds. If the time between two measurements is smaller than shoulder, the time stamps are duplicated.

drop

logical; if drop = TRUE (default), time stamps without an assigned UniqueID are removed from the final data frame.

Value

A merged data.frame between the inputfile and the auxfile, with the additional columns flag, Etime and obs.length.

Details

Note that, if the time between two measurements is smaller than shoulder, the time stamps are duplicated, which can result in an output data frame larger (more rows) than the inputfile. Conversely, if drop = TRUE, time stamps without an assigned UniqueID are removed from the final data frame, which can result in an output data frame smaller (less rows) than the inputfile.

See also

Examples

# How to use in multiple situations:
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

## with a LGR instrument and an auxiliary file (.txt)
aux.path <- system.file("extdata", "aux_UGGA/aux_UGGA.txt", package = "goFlux")
auxfile <- read.delim(aux.path) %>%
  mutate(start.time = as.POSIXct(start.time, tz = "UTC"))
data(imp.UGGA)
autoID.UGGA <- autoID(inputfile = imp.UGGA, auxfile = auxfile, obs.length = 180)

## with a LI-COR instrument and the Smart Chamber as auxiliary file
data(imp.LI8200)
data(imp.LI7820)
autoID.LI7820 <- autoID(inputfile = imp.LI7820, deadband = 20, auxfile = imp.LI8200)

## with the LI-6400 and no auxiliary file
data(imp.LI6400)
ow.LI6400 <- obs.win(inputfile = imp.LI6400, shoulder = 0)