Looped manual identification of start and end of gas measurements

DEPRECATED: This function will be removed in a future version of the package. Use the function click.peak2 instead.

Identify the start and the end of a measurement by clicking on them in a scatter plot. To use in a loop with multiple measurements, first use the function obs.win to identify the observation window of each measurement and then use the wrapper function click.peak.loop with lapply (see example below).

click.peak.loop(
  x,
  flux.unique,
  gastype = "CO2dry_ppm",
  sleep = 3,
  plot.lim = c(380, 1000),
  warn.length = 60
)

Arguments

x

a numerical sequence that indicates objects in a list. Must be used with lapply after the function obs.win. See examples below to see how to use with lapply.

flux.unique

data.frame; output from the function obs.win. Must contain the columns gastype (see below), POSIX.time and UniqueID.

gastype

character string; specifies which gas should be displayed on the plot to manually select start time and end time of measurements. Must be one of the following: "CO2dry_ppm", "COdry_ppb", "CH4dry_ppb", "N2Odry_ppb", "NH3dry_ppb" or "H2O_ppm". Default is "CO2dry_ppm".

sleep

numerical value; delay before closing the resulting plot. When used with the function click.peak.loop, grants a delay between measurements to visually inspect the output before processing the next measurement. Sleep must be shorter than 10 seconds.

plot.lim

numerical vector of length 2; sets the Y axis limits in the plots. Default values are set for a typical gas measurement of "CO2dry_ppm" from soils: plot.lim = c(380,1000).

warn.length

numerical value; limit under which a measurement is flagged for being too short (nb.obs < warn.length). Default value is warn.length = 60.

Value

A list of data.frame, split by UniqueID, identical to the input flux.unique, with the additional columns flag, Etime, start.time_corr, end.time_corr and obs.length_corr.

Details

The argument plot.lim is used to remove any data points below and above the plot limits for a better view of the scatter plot. If the gas measurements are larger than the minimum or smaller than the maximum plot limit values, then the plot will automatically zoom in and adjust to those values. The default plot limits are set for a typical gas measurement of "CO2dry_ppm" from a soil respiration measurement: plot.lim = c(380,1000), where 380 ppm is the minimum plotted concentration, which should be close to atmospheric concentration, and 1000 ppm is the maximum plotted concentration, which correspond to a maximal accumulated concentration in the chamber before considering it an outlier (e.g. caused by breath or gas bubble). For other gasses, the user must specify the plot limits themselves. Here are some suggestions of plot limits for the other gases:

"CH4dry_ppb": plot.lim = c(2200, 1800)

"N2Odry_ppb": plot.lim = c(250, 500)

"NH3dry_ppb": plot.lim = c(0, 200)

"COdry_ppb": plot.lim = c(0, 200)

"H2O_ppm": plot.lim = c(10000, 20000)

These values will vary depending on ecosystem type and chamber application scheme.

warn.length is the limit below which the chamber closure time is flagged for being too short (nb.obs < warn.length). Portable greenhouse gas analyzers typically measure at a frequency of 1 Hz. Therefore, for the default setting of warn.length = 60, the chamber closure time should be approximately one minute (60 seconds). If the number of observations is smaller than the threshold, a warning is printed after the loop: "Number of observations for UniqueID: 'UniqueID' is X observations".

In gastype, the gas species listed are the ones for which this package has been adapted. Please write to the maintainer of this package for adaptation of additional gases.

See also