Manual editing of observation window of flux measurements

Crop measurements from the start, from the end, or adjust the deadband. Individual parameters can be provided for each measurement, in an auxiliary file, or the same parameters can be applied to all measurements.

crop.meas(
  dataframe,
  auxfile = NULL,
  crop_start = 0,
  crop_end = 0,
  deadband = 0,
  max.obs.length = NULL,
  cham.close = "start.time",
  cham.open = "end.time"
)

Arguments

dataframe

a dataframe containing gas measurements, output from import or ID functions (click.peak2 or manID). Must contain the columns UniqueID, start.time, end.time and Etime.

auxfile

data.frame; auxiliary data frame containing the columns UniqueID, crop_start, crop_end, deadband and/or max.obs.length.

crop_start

numerical; number of observations to remove at the start of the measurement. To provide a unique value per UniqueID, the column crop_start must be present in the auxfile and the argument must be set to crop_start = "aux".

crop_end

numerical; number of observations to remove at the end of the measurement. To provide a unique value per UniqueID, the column crop_end must be present in the auxfile and the argument must be set to crop_end = "aux".

deadband

numerical; define a deadband at the start of measurements (seconds). To provide a unique value per UniqueID, the column deadband must be present in the auxfile and the argument must be set to deadband = "aux".

max.obs.length

numerical; maximum length of measurement (seconds) to keep after cham.close. To provide a unique value per UniqueID, the column max.obs.length must be present in the auxfile and the argument must be set to max.obs.length = "aux".

cham.close

character; if cham.close = "start.time" (default), the original start.time before cropping will be saved into a new column cham.close. Set cham.close = NULL to drop the original start.time.

cham.open

character; if cham.open = "end.time" (default), the original end.time before cropping will be saved into a new column cham.open. Set cham.open = NULL to drop the original end.time.

Value

a data frame, with corrected Etime and flag, based on newly defined start.time, end.time and deadband.

Details

Note that crop_start and deadband "crop" the start of the measurements in the same way.

Note that if max.obs.length is defined, end.time is calculated based on cham.close (original start.time in dataframe, if cham.close = "start.time") and max.obs.length, regardless of other defined values in the function. If cham.close = NULL, end.time is calculated based on start.time and max.obs.length.

If cham.close and cham.open are originally present in dataframe, set arguments cham.close and cham.open to NULL.

Note that the arguments crop_start and crop_end accept negative values and can be used to include more data points before and after the previously defined start.time and end.time.

Examples

# Load example data
data(manID.G2201i)

# Create an auxfile to define different crop_start and crop_end per UniqueID
auxfile <- cbind.data.frame(UniqueID = c("meas1", "meas2"),
                            crop_start = c(2,10),
                            crop_end = c(5,0))

# The arguments "crop_start" and "crop_end" must be set to "aux" to use
# unique values from the auxfile
crop.G2201i <- crop.meas(dataframe = manID.G2201i, auxfile = auxfile,
                         crop_start = "aux", crop_end = "aux")

# Alternatively, cropping can be applied equally across all measurements
crop.G2201i <- crop.meas(dataframe = manID.G2201i, auxfile = auxfile,
                         crop_start = 5, crop_end = 5)