Import function for the Picarro G2201-i Isotopic Analyzer

Imports single raw gas measurement files from the Picarro G2201-i with the extension .dat (CO2 and CH4 isotopic analyzer)

import.G2201i(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  save = FALSE,
  keep_all = FALSE,
  prec = c(0.2, 50, 100),
  CH4 = "HP_12CH4_dry",
  CO2 = "12CO2_dry",
  sum = TRUE,
  range = 10
)

Arguments

inputfile

character string; the name of a file with the extension .dat

date.format

character string; specifies the date format found in the raw data file. Choose one of the following: "dmy", "ymd", or "mdy". Default is "ymd", as it is the date format from the example data file provided.

timezone

character string; a time zone in which to import the data to POSIXct format. Default is "UTC". Note about time zone: it is recommended to use the time zone "UTC" to avoid any issue related to summer time and winter time changes.

save

logical; if save = TRUE, saves the file as an .RData file in a RData folder in the current working directory. If save = FALSE, returns the file in the Console, or load in the Environment if assigned to an object.

keep_all

logical; if keep_all = TRUE, keep all columns from the raw file. The default is keep_all = FALSE, and columns that are not necessary for gas flux calculation are removed.

prec

numerical vector; the precision of the instrument for each gas, in the following order: "CO2dry_ppm", "CH4dry_ppb" and "H2O_ppm". The default is prec = c(0.2, 50, 100).

CH4

character string; which column should be used for CH4 concentration. By default, CH4 = "HP_12CH4_dry".

CO2

character string; which column should be used for CO2 concentration. By default, CO2 = "12CO2_dry".

sum

logical; if sum = TRUE (default), the arguments CH4 and CO2 are ignored and, instead, both fractions (carbon isotope 12 and 13) of each gas, are summed after correction for water vapor, to obtain the total gas concentration. See the argument range for the selection between the HP or the HR measurement of CH4 concentration.

range

numerical value; at which threshold should the function select between high precision (HP) and high resolution (HR) of CH4 concentration? The default is range = 10. If the methane concentration recorded in the column HP_12CH4_dry is above 10 ppm, the concentration from the columns HR_12CH4_dry and HR_13CH4 are used for the sum instead of HP.

Value

A data frame containing raw data from Picarro G2201-i isotopic analyzer.

Details

In date.format, the date format refers to a date found in the raw data file, not the date format in the file name. For the instrument G2201-i, the date is found in the column "DATE".

The arguments CH4 and CO2 require the user to specify which columns contain the two gases of interest among the following for CH4:

HP_12CH4

HP_12CH4_dry

HP_13CH4

HR_12CH4

HR_12CH4_dry

HR_13CH4

And among the following for CO2:

12CO2

12CO2_dry

13CO2

13CO2_dry

Note that, HP stands for CH4 concentration measured with strong line for high precision but limited range and HR stands for CH4 concentration measured with high range (between 10 and 1000 ppm). In addition, _dry stands for dry mole fraction of gases, corrected for water vapor. Finally, 12 and 13 indicate each carbon isotope contained in each gas. If one selects one of the non-dry mole fraction, it will be corrected for water vapor during the import process.

Note that this function was designed for the following units in the raw file:

ppm for CO2 and CH4

percent (%) for H2O

Torr for pressure

Celsius for temperature

If your Picarro G2201-i uses different units, either convert the units after import, change the settings on your instrument, or contact the maintainer of this package for support.

The precision of the instrument is needed to restrict kappa-max (k.max) in the non-linear flux calculation (HM.flux). Kappa-max is inversely proportional to instrument precision. If the precision of your instrument is unknown, it is better to use a low value (e.g. 1 ppm) to allow for more curvature, especially for water vapor fluxes, or very long measurements, that are normally curved. The default values given for instrument precision are the ones found online for the latest model of this instrument, available at the time of the creation of this function (05-2025).

See also

Examples

# Load file from downloaded package
file.path <- system.file("extdata", "G2201i/2025/05/13/G2201i.dat", package = "goFlux")

# Run function
imp.G2201i <- import.G2201i(inputfile = file.path)