Import function for the SkyLine2D system linked to one GHG analyzer

Imports single raw gas measurement files from the SkyLine2D system from Earth Bound Scientific from the datalog file created by GAIA2TECH. Default settings are set for the Picarro G2508 (CO2, CH4, N2O and H2O), to match the example data file provided with the package.

import.skyline(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  save = FALSE,
  keep_all = FALSE,
  deadband = 0,
  shoulder = 20,
  CH.col = "CH ID",
  CH.clo.col = "Chamber closed",
  sensor1 = "Analog Sensor1",
  sensor2 = "Analog Sensor2",
  inst = "Instrument",
  gas = c("CO2", "CH4 Dry", "N2O Dry", "H2O"),
  prec = c(0.24, 0.3, 5, 500),
  dry = T
)

Arguments

inputfile

character string; the name of a file with the extension .csv

date.format

character string; specifies the date format found in the raw data file. Choose one of the following: "dmy", "ymd", or "mdy". Default is "ymd", as it is the date format from the example data file provided.

timezone

character string; a time zone in which to import the data to POSIXct format. Default is "UTC". Note about time zone: it is recommended to use the time zone "UTC" to avoid any issue related to summer time and winter time changes.

save

logical; if save = TRUE, saves the file as an .RData file in a RData folder in the current working directory. If save = FALSE, returns the file in the Console, or load in the Environment if assigned to an object.

keep_all

logical; if keep_all = TRUE, keep all columns from the raw file. The default is keep_all = FALSE, and columns that are not necessary for gas flux calculation are removed.

deadband

numerical; define a deadband at the start of measurements (seconds).

shoulder

numerical; include background data points before and after the measurement (seconds).

CH.clo.col, CH.col

character string; a pattern to match columns that fit the corresponding parameters: CH.col = "CH ID" and CH.clo.col = "Chamber closed".

sensor1, sensor2

character strings; a pattern to match the columns containing the name of each sensor. By default, sensor1 = "Analog Sensor1", sensor2 = "Analog Sensor2", to match the example data file provided with the package. Set to NULL to ignore these columns.

inst

character string; a pattern to match the column containing the name of the GHG analyzer: inst = "Instrument".

gas

character vector; a pattern to match the columns containing each gas measurement. By default, gas = c("CO2", "CH4 Dry", "N2O Dry", "H2O"), to match the example data file provided with the package.

prec

numerical vector; the precision of the instrument for each gas mentioned in gas. By default, prec = c(0.24, 0.3, 5, 500), to match the example data file provided with the package. Note that the order in the arguments prec must match the order of the arguments in gas.

dry

logical; are the gas measurements compensated for water vapor (dry fraction)? If dry = TRUE (default), the gases are already compensated for water vapor. Otherwise, the gases will be converted to a dry fraction.

Value

A data frame containing raw data from the SkyLine2D system

Details

This function has been designed for the SkyLine2D system file generated by GAIA2TECH synced with the Picarro G2508 (CO2, CH4, N2O and H2O)). If this function could be useful for you, but does not meet your needs, please contact the maintainer of this package for potential adaptations.

In date.format, the date format refers to a date found in the raw data file, not the date format in the file name. For the SkyLine2D system, the date is found in the column "Titles:". Be mindful that when you open a .csv file in Excel, the software often changes date formats. To avoid any discrepancy, open the file in a text file.

The argument CH.clo.col corresponds to the column "Chamber closed". Any number above 0 in that column indicate that the chamber is closed. The argument CH.col corresponds to the column "CH ID" indicating which chamber is active.

For each gas, the units are taken from the second row in the raw file (Units:). If your instrument uses different units, either convert the units after import, change the settings on your instruments, or contact the maintainer of this package for support.

Regarding the parameters dry, in case of uncertainty, either contact your technical support or assume that gases are compensated for water vapor, which is normally the case.

As opposed to the other import functions, there is no option to "keep_all" with this instrument. If you would like to import additional data using this function, please contact the maintainer of this package for support.

The precision of the instrument is needed to restrict kappa-max (k.max) in the non-linear flux calculation (HM.flux). Kappa-max is inversely proportional to instrument precision. If the precision of your instrument is unknown, it is better to use a low value (e.g. 1 ppm for CO2 and H2O, or 1 ppb for CH4 and N2O) to allow for more curvature, especially for water vapor fluxes, or very long measurements, that are normally curved. The default values given for instrument precision are the ones found online for the latest models of the Picarro G2508, available at the time of the creation of this function (10-2025).

See also

Examples

# Load file from downloaded package
file.path <- system.file("extdata", "skyline/skyline.csv", package = "goFlux")

# Run function
imp.skyline <- import.skyline(inputfile = file.path)