Import function for the automated chamber ECOFlux (GAIA2TECH) synced with multiple chambers and instruments

Imports single raw gas measurement files from the automated chamber ECOFlux (GAIA2TECH) linked with up to three greenhouse gas analyzers. Default settings are set for two LI-COR instruments: LI-7810 (CO2, CH4 and H2O) and LI7820 (N2O and H2O), to match the example data file provided with the package.

GAIA_import(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  background = FALSE,
  save = FALSE,
  CH.col = "COM5A0",
  SWC.col = "1C08_Soil Moisture",
  Tsoil.col = "1C07_Soil Temperature",
  Tcham.col = "2C07_Chamber Temperature",
  PAR.col = "3C07_Sunlight",
  WTD.col = NULL,
  Op.stat.col = "0C06_OperatingStatus",
  inst1 = "XT2C00_Instrument",
  inst2 = "XT3C00_Instrument",
  inst3 = NULL,
  gas1 = c("XT2C04_CH4", "XT2C05_CO2", "XT2C06_H2O"),
  gas2 = c("XT3C04_N2O", "XT3C05_H2O"),
  gas3 = NULL,
  prec1 = c(0.6, 3.5, 45),
  prec2 = c(0.4, 45),
  prec3 = NULL,
  dry1 = T,
  dry2 = T,
  dry3 = NULL,
  manual = FALSE,
  sep = "\t",
  skip = 1
)

import.GAIA(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  background = FALSE,
  save = FALSE,
  CH.col = "COM5A0",
  SWC.col = "1C08_Soil Moisture",
  Tsoil.col = "1C07_Soil Temperature",
  Tcham.col = "2C07_Chamber Temperature",
  PAR.col = "3C07_Sunlight",
  WTD.col = NULL,
  Op.stat.col = "0C06_OperatingStatus",
  inst1 = "XT2C00_Instrument",
  inst2 = "XT3C00_Instrument",
  inst3 = NULL,
  gas1 = c("XT2C04_CH4", "XT2C05_CO2", "XT2C06_H2O"),
  gas2 = c("XT3C04_N2O", "XT3C05_H2O"),
  gas3 = NULL,
  prec1 = c(0.6, 3.5, 45),
  prec2 = c(0.4, 45),
  prec3 = NULL,
  dry1 = T,
  dry2 = T,
  dry3 = NULL,
  manual = FALSE,
  sep = "\t",
  skip = 1
)

Arguments

inputfile

character string; the name of a file with the extension .csv

date.format

character string; specifies the date format found in the raw data file. Choose one of the following: "dmy", "ymd", or "mdy". Default is "ymd", as it is the date format from the example data file provided.

timezone

character string; a time zone in which to import the data to POSIXct format. Default is "UTC". Note about time zone: it is recommended to use the time zone "UTC" to avoid any issue related to summer time and winter time changes.

pivot

character string; either "long" or "wide". If pivot = "long", each column containing information about Tsoil, Tcham, SWC, PAR and operating status (Op.stat) will be saved in a single column per parameter. If pivot = "wide", the default display of one column per chamber per parameter will be used.

active

logical; if active = TRUE, preserve data for active chambers only.

flag

numeric vector; indicates the operating status that should be used for the flux calculation. Default is flag = c(7,11), where 7 indicates "Chamber Idle Closed Clear" and 11 indicates "Chamber Idle Closed Dark".

background

logical; if background = FALSE, removes all data from activ.cham == "Background".

save

logical; if save = TRUE, saves the file as an .RData file in a RData folder in the current working directory. If save = FALSE, returns the file in the Console, or load in the Environment if assigned to an object.

PAR.col, Tcham.col, Tsoil.col, SWC.col, WTD.col

character vector; a vector of patterns to match all columns that fit the corresponding parameters. See details below for more information.

Op.stat.col, CH.col

character string; a pattern to match columns that fit the corresponding parameters: see details below.

inst1, inst2, inst3

character strings; a pattern to match the columns containing the name of each instrument. By default, inst1 = "XT2C00_Instrument", inst2 = "XT3C00_Instrument", and inst3 = NULL, to match the example data file provided with the package.

gas1, gas2, gas3

character vectors; a pattern to match the columns containing each gas measurement. By default, gas1 = c("XT2C04_CH4", "XT2C05_CO2", "XT2C06_H2O"), gas2 = c("XT3C04_N2O", "XT3C05_H2O"), and gas3 = NULL, to match the example data file provided with the package.

prec1, prec2, prec3

numerical vectors; the precision of the instrument for each gas mentioned in gas1, gas2 and gas3. By default, prec1 = c(0.6, 3.5, 45), prec2 = c(0.4, 45), and prec3 = NULL, to match the example data file provided with the package. Note that the order in the arguments precX must match the order of the arguments in gasX.

dry1, dry2, dry3

logical; are the gas measurements compensated for water vapor (dry fraction)? If dryX = TRUE (default), the gases are compensated for water vapor, and will be named accordingly. For example, the column "XT2C05_CO2" will become "CO2dry_ppm".

manual

logical; indicates if measurements were made manually and there are no records of chamber identification, sequence, etc. If TRUE, the arguments Op.stat.col, CH.col, active, background, flag, and pivot are ignored.

sep

character string defining the field separator character. Values on each line of the file are separated by this character. By default, sep = "\t" for tabulation.

skip

integer; the number of lines of the data file to skip before beginning to read data. By default, skip = 1.

Value

A data frame containing raw data from the automated chamber ECOFlux (GAIA2TECH)

Details

This function has been designed for the automated chamber ECOFlux (GAIA2TECH) synced with up to 9 chambers and two GHG analyzers from LI-COR (LI-7810: CO2, CH4 and H2O / LI7820: N2O and H2O). If this function could be useful for you, but does not meet your needs, please contact the maintainer of this package for potential adaptations.

In date.format, the date format refers to a date found in the raw data file, not the date format in the file name. For the automated chamber ECOFlux, the date is found in the column "Titles:".

The arguments PAR.col, Tcham.col, Tsoil.col, SWC.col, and WTD.col correspond to different types of probes linked to the ECOFlux chamber: PAR (sunlight), chamber temperature, soil temperature, soil water content and water table depth, respectively. All columns containing the patterns specified in these arguments will be selected and renamed with the corresponding argument. For example, if PAR = "3C07_Sunlight", all columns containing the pattern "3C07_Sunlight" will be renamed with the pattern "_PAR". Then, if pivot = "long", all columns with the pattern "_PAR" will be merged together. If PAR = c("3C07_Sunlight1", "3C07_Sunlight2"), these patterns will be renamed to PAR1 and PAR2.

The argument Op.stat.col corresponds to the columns Operating Status of each chamber. The argument CH.col indicates a character string preceding the chamber number for each column of the raw data. For example, the column "COM5A010C06_OperatingStatus" in the raw data file contains the Operating Status for the chamber 1 if CH.col = "COM5A0" and Op.stat.col = "0C06_OperatingStatus". If these columns are absent from the raw data, set arguments to NULL.

Note that this function was designed for the following units in the raw file:

kPa for atmospheric pressure

volumetric water content (%) for soil moisture

Celsius for air temperature

µmol photons m-2s-1 for PAR

For each gas, the units are taken from the second row in the raw file (Units:). If your instruments use different units, either convert the units after import, change the settings on your instruments, or contact the maintainer of this package for support.

Regarding the parameters dryX, in case of uncertainty, either contact your technical support or assume that gases are compensated for water vapor, which is normally the case.

As opposed to the other import functions, there is no option to "keep_all" with this instrument. If you would like to import additional data using this function, please contact the maintainer of this package for support.

The precision of the instrument is needed to restrict kappa-max (k.max) in the non-linear flux calculation (HM.flux). Kappa-max is inversely proportional to instrument precision. If the precision of your instrument is unknown, it is better to use a low value (e.g. 1 ppm for CO2 and H2O, or 1 ppb for CH4 and N2O) to allow for more curvature, especially for water vapor fluxes, or very long measurements, that are normally curved. The default values given for instrument precision are the ones found online for the latest models of the LI-7810 and LI-7820, available at the time of the creation of this function (11-2023).

See also

Examples

# Load file from downloaded package
file.path <- system.file("extdata", "GAIA/GAIA.csv", package = "goFlux")

# Run function
imp.GAIA <- import.GAIA(inputfile = file.path)
#> Warning: The same gas was measured by two instruments. The suffix 'inst1', 'inst2' or 'inst3' was added to each duplicated gas column to avoid confusion. 'inst1' = LICOR78XX#1, 'inst2' = LICOR78XX#2 and 'inst3' = NULL