Looping function for quick import of multiple raw gas measurement files

Imports all raw gas measurement files contained in a folder. Adapted to multiple greenhouse gas analyzers and other instruments:

LI-COR: LI-6400, LI-7810, LI-7820, LI-8100, LI-8200 (smart chamber)

Los Gatos Research (LGR): ultra-portable GGA (GLA132 series), micro ultra-portable GGA (GLA131 series) and N2OM1 (GLA151 series)

GAIA2TECH (DMR) automated chamber ECOFlux

Picarro: G2508 and G4301

Gasmet: DX4015

PP-Systems: EGM-5

import2RData(
  path,
  instrument,
  date.format,
  timezone = "UTC",
  keep_all,
  prec,
  proc.data.field = NULL
)

Arguments

path

character string; a folder path containing all files to be imported. Ideally, to avoid any errors, the folder should only contain the files to be imported.

instrument

character string; specifies which instrument was used to generate the files contained in the folder path. Choose one of the following: "DX4015", "EGM5", "G2508", "G4301" "GAIA", "LI-6400", "LI-7810", "LI-7820", "LI-8100", "LI-8200", "LGR" or "N2OM1". For more information about an instrument, see the section "See Also" below.

date.format

character string; specifies the date format found in the raw data file. Choose one of the following: "dmy", "ymd", or "mdy".

timezone

character string; a time zone in which to import the data to POSIXct format. Default is "UTC". Note about time zone: it is recommended to use the time zone "UTC" to avoid any issue related to summer time and winter time changes.

keep_all

logical; if keep_all = TRUE, keep all columns from the raw file. The default is keep_all = FALSE, and columns that are not necessary for gas flux calculation are removed.

prec

numerical vector; the precision of the instrument for each gas. Look at the examples below, or the help for each import function of each instrument, to know what values to use.

proc.data.field

numeric value; select the process data field used with the EGM-5. There are 5 different modes available: (1) Measure mode; (2) SRC or Custom mode; (3) CPY mode; (4) Injection mode; or (5) Static mode. If no value specified, the parameters will be named "param1", "param2", "param3", "param4" and "param5".

Value

A data frame saved as RData in a newly created folder, RData, into the working directory.

Details

In date.format, the date format refers to a date found in the raw data file, not the date format in the file name. For most instruments, the date is one of the columns in the raw data file:

DX4015: column Date

EGM5: column Date

G2508: column DATE

G4301: column DATE

GAIA: column Titles:

LGR: column Time

LI-6400: (see comment below)

LI-7810: column DATE

LI-7820: column DATE

LI-8100: column Date

LI-8200: (see comment below)

N2OM1: column Time

For the instrument LI-6400, the date is found in one of the first lines in a format containing abbreviations, for example "Thr Aug 6 2020", which would be the date format "mdy". For the instrument LI-8200, the date is found under one of the measurements, next to '"Date":'.

The argument proc.data.field is used with the PP-Sytems EGM-5 instrument only. Look up the EGM-5 Operation Manual at page 90 for more details about the different Process Data Fields (proc.data.field).

See also

Examples

# Examples on how to use it with all instruments.
# The default time zone "UTC" is used in all cases.

# with the Gasmet instrument DX4015
file.path <- system.file("extdata/DX4015", package = "goFlux")
import2RData(path = file.path, instrument = "DX4015",
             date.format = "ymd", keep_all = FALSE,
             prec = c(1.6, 23, 13, 2, 23, 33))
#> DX4015 saved as DX4015_DX4015_imp.RData in RData folder, in working directory

# with the PP-Systems EGM-5
# The argument proc.data.field is used with the EGM5 instrument only.
file.path <- system.file("extdata/EGM5", package = "goFlux")
import2RData(path = file.path, instrument = "EGM5",
             date.format = "dmy", keep_all = FALSE,
             prec = c(3, 1, 500), proc.data.field = 2)
#> Warning: There were 15 warnings in `mutate()`.
#> The first warning was:
#> ℹ In argument: `across(Plot_No:ncol(.), as.numeric)`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 14 remaining warnings.
#> EGM5 saved as EGM5_EGM5_imp.RData in RData folder, in working directory

# with the Picarro instrument G2508
file.path <- system.file("extdata/G2508", package = "goFlux")
import2RData(path = file.path, instrument = "G2508",
             date.format = "ymd", keep_all = FALSE,
             prec = c(0.24, 0.3, 5, 0.16, 500))
#> G2508 saved as G2508_G2508_imp.RData in RData folder, in working directory

# with the Picarro instrument G4301
file.path <- system.file("extdata/G4301", package = "goFlux")
import2RData(path = file.path, instrument = "G4301",
             date.format = "ymd", keep_all = FALSE,
             prec = c(0.025, 0.1, 10))
#> G4301 saved as G4301_G4301_imp.RData in RData folder, in working directory

# with the automated chamber ECOFlux (GAIA2TECH)
# with this instrument, "keep_all" is not a valid argument.
file.path <- system.file("extdata/GAIA", package = "goFlux")
import2RData(path = file.path, instrument = "GAIA",
             date.format = "ymd", prec = c(3.5, 0.6, 0.4, 45, 45))
#> GAIA saved as GAIA_GAIA_imp.RData in RData folder, in working directory

# with Los Gatos Research UGGA (GLA132 series)
file.path <- system.file("extdata/LGR", package = "goFlux")
import2RData(path = file.path, instrument = "LGR",
             date.format = "dmy", keep_all = FALSE,
             prec = c(0.2, 1.4, 50))
#> LGR saved as LGR_LGR_imp.RData in RData folder, in working directory

# with Los Gatos Research m-GGA (GLA131 series)
file.path <- system.file("extdata/LGR", package = "goFlux")
import2RData(path = file.path, instrument = "LGR",
             date.format = "dmy", keep_all = FALSE,
             prec = c(0.35, 0.9, 200))
#> LGR saved as LGR_LGR_imp.RData in RData folder, in working directory

# with Los Gatos Research N2OM1 (GLA151 series)
file.path <- system.file("extdata/N2OM1", package = "goFlux")
import2RData(path = file.path, instrument = "N2OM1",
             date.format = "dmy", keep_all = FALSE,
             prec = c(0.5, 2, 50))
#> N2OM1 saved as N2OM1_N2OM1_imp.RData in RData folder, in working directory

# with the LI-COR LI-6400 gas analyzer
file.path <- system.file("extdata/LI6400", package = "goFlux")
import2RData(path = file.path, instrument = "LI-6400",
             date.format = "mdy", keep_all = FALSE,
             prec = c(0.15, 20))
#> LI6400 saved as LI6400_LI6400_imp.RData in RData folder, in working directory

# with the LI-COR LI-7810 gas analyzer
file.path <- system.file("extdata/LI7810", package = "goFlux")
import2RData(path = file.path, instrument = "LI-7810",
             date.format = "ymd", keep_all = FALSE,
             prec = c(3.5, 0.6, 45))
#> LI7810 saved as LI7810_LI7810_imp.RData in RData folder, in working directory

# with the LI-COR LI-7820 gas analyzer
file.path <- system.file("extdata/LI7820", package = "goFlux")
import2RData(path = file.path, instrument = "LI-7820",
             date.format = "ymd", keep_all = FALSE,
             prec = c(0.4, 45))
#> LI7820 saved as LI7820_LI7820_imp.RData in RData folder, in working directory

# with the LI-COR LI-8100 gas analyzer
file.path <- system.file("extdata/LI8100", package = "goFlux")
import2RData(path = file.path, instrument = "LI-8100",
             date.format = "ymd", keep_all = FALSE,
             prec = c(1, 10))
#> LI8100 saved as LI8100_LI8100_imp.RData in RData folder, in working directory

# with the LI-COR smart chamber (LI-8200)
# with this instrument, "keep_all" and "prec" are not valid arguments.
file.path <- system.file("extdata/LI8200", package = "goFlux")
import2RData(path = file.path, instrument = "LI-8200",
             date.format = "ymd")
#> LI8200 saved as LI8200_LI8200_imp.RData in RData folder, in working directory