Looping function for quick import of multiple raw gas measurement files

Imports all raw gas measurement files contained in a folder. Adapted to multiple greenhouse gas analyzers and other instruments:

LI-COR: LI-6400, LI-7810, LI-7820, LI-8100, LI-8200 (smart chamber), LI-8250 (Multiplexer)

Los Gatos Research (LGR): ultra-portable GGA (GLA132 series), micro ultra-portable GGA (GLA131 series), N2OM1 (GLA151 series), and the N2O isotopic analyzer N2Oi2 (GLA451 series)

GAIA2TECH (DMR): automated chamber ECOFlux

Picarro: G2201-i, G2508 and G4301

Gasmet: DX4015 and GT5000

PP-Systems: EGM-5

Aeris Technologies: N2O/ CO2 high accuracy analyzer and CH4/ C2H6 natural gas detection system

Earth Bound Scientific: SkyLine2D system

GasmetPD: Custom multiplexer from the University of Padova, Italy

import2RData(
  path,
  instrument,
  date.format,
  timezone = "UTC",
  keep_all = FALSE,
  prec,
  merge = FALSE,
  proc.data.field = NULL,
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  background = FALSE,
  deadband = 0,
  shoulder = 20,
  SWC.col,
  Tsoil.col,
  Tcham.col,
  PAR.col,
  WTD.col,
  CH.col,
  Op.stat.col,
  CH.clo.col,
  sensor1,
  sensor2,
  inst1,
  inst2 = NULL,
  inst3 = NULL,
  inst,
  gas1,
  gas2 = NULL,
  gas3 = NULL,
  gas,
  prec1,
  prec2 = NULL,
  prec3 = NULL,
  dry1 = T,
  dry2 = T,
  dry3 = NULL,
  dry = T,
  manual = F,
  sep = "\t",
  skip = 1,
  CH4 = "HP_12CH4_dry",
  CO2 = "12CO2_dry",
  sum = TRUE,
  range = 10
)

Arguments

path

character string; a folder path containing all files to be imported. Ideally, to avoid any errors, the folder should only contain the files to be imported.

instrument

character string; specifies which instrument was used to generate the files contained in the folder path. Choose one of the following: "DX4015", "EGM5", "G2201i", "G2508", "G4301", "GAIA", "GasmetPD", "GT5000", "LI-6400", "LI-7810", "LI-7820", "LI-8100", "LI-8200", "LI-8250", "N2OM1", "N2Oi2", "skyline", "uCH4", "uN2O", "UGGA". For more information about an instrument, see the section "See Also" below.

date.format

character string; specifies the date format found in the raw data file. Choose one of the following: "dmy", "ymd", or "mdy".

timezone

character string; a time zone in which to import the data to POSIXct format. Default is "UTC". Note about time zone: it is recommended to use the time zone "UTC" to avoid any issue related to summer time and winter time changes.

keep_all

logical; if keep_all = TRUE, keep all columns from the raw file. The default is keep_all = FALSE, and columns that are not necessary for gas flux calculation are removed.

prec, prec1, prec2, prec3

numerical vector; the precision of the instrument for each gas. Look at the examples below, or the help for each import function of each instrument, to know what values to use.

merge

logical; if TRUE, merge all files as one data frame and return it in the Console, or load it in the Environment if assigned to an object.

proc.data.field

numeric value; select the process data field used with the EGM-5. See import.EGM5 for more details.

pivot

character string; either "long" or "wide". If pivot = "long", each column containing information about Tsoil, Tcham, SWC, PAR and operating status (Op.stat) will be saved in a single column per parameter. If pivot = "wide", the default display of one column per chamber per parameter will be used. Used with instrument = "GAIA" only.

active

logical; if active = TRUE, preserve data for active chambers only. Used with instrument = "GAIA" only.

flag

numeric vector; indicates the operating status that should be used for the flux calculation. Default is flag = c(7,11), where 7 indicates "Chamber Idle Closed Clear" and 11 indicates "Chamber Idle Closed Dark". Used with instrument = "GAIA" only.

background

logical; if background = FALSE, removes all data from opened chambers. See import.GAIA for more details.

deadband

numerical; define a deadband at the start of measurements (seconds). Used with instrument = "skyline" only.

shoulder

numerical; include background data points before and after the measurement (seconds). Used with instrument = "skyline" only.

CH.clo.col, Op.stat.col, PAR.col, Tcham.col, Tsoil.col, SWC.col, WTD.col, CH.col

character vector; a pattern to match the columns that fit the corresponding parameter. See import.GAIA, import.LI8250, or import.skyline for more details.

sensor1, sensor2

character strings; a pattern to match the columns containing the name of each sensor. By default, sensor1 = "Analog Sensor1", sensor2 = "Analog Sensor2", to match the example data file provided with the package. Set to NULL to ignore these columns. Used with instrument = "skyline" only.

inst, inst1, inst2, inst3

character strings; a pattern to match the columns containing the name of each instrument. See import.GAIA, import.LI8250 or instrument = "skyline" for more details.

gas, gas1, gas2, gas3

character vectors; a pattern to match the columns containing each gas measurement. See import.GAIA, import.LI8250, import.GasmetPD or import.skyline for more details.

dry, dry1, dry2, dry3

logical; are the gas measurements compensated for water vapor (dry fraction)? If dryX = TRUE (default), the gases are compensated for water vapor, and will be named accordingly. See import.GAIA, import.LI8250 or import.skyline for more details.

manual

logical; indicates if measurements were made manually and there are no records of chamber identification, sequence, etc. If TRUE, the arguments Op.stat.col, CH.col, active, background, flag, and pivot are ignored. Used with instrument = "GAIA" only.

sep

character string defining the field separator character. Values on each line of the file are separated by this character. By default, sep = "\t" for tabulation. Used with instrument = "GAIA" only.

skip

integer; the number of lines of the data file to skip before beginning to read data. By default, skip = 1. Used with instrument = "GAIA" only.

CH4

character string; which column should be used for CH4 concentration. By default, CH4 = "HP_12CH4_dry". Used with instrument = "G2201i" only.

CO2

character string; which column should be used for CO2 concentration. By default, CO2 = "12CO2_dry".Used with instrument = "G2201i" only.

sum

logical; if sum = TRUE (default), the arguments CH4 and CO2 are ignored and, instead, both fractions (carbon isotope 12 and 13) of each gas, are summed after correction for water vapor, to obtain the total gas concentration. See the argument range for the selection between the HP or the HR measurement of CH4 concentration. Used with instrument = "G2201i" only.

range

numerical value; at which threshold should the function select between high precision (HP) and high resolution (HR) of CH4 concentration? The default is range = 10. If the methane concentration recorded in the column HP_12CH4_dry is above 10 ppm, the concentration from the columns HR_12CH4_dry and HR_13CH4 are used for the sum instead of HP. Used with instrument = "G2201i" only.

Value

Raw files saved as RData in a newly created folder, RData, into the working directory.

Details

In date.format, the date format refers to a date found in the raw data file, not the date format in the file name. For most instruments, the date is one of the columns in the raw data file:

DX4015: column Date

EGM5: column Date

G2201i: column DATE

G2508: column DATE

G4301: column DATE

GAIA: column Titles:

GasmetPD: column Date

GT5000: column Date

LI-6400: (see comment below)

LI-7810: column DATE

LI-7820: column DATE

LI-8100: column Date

LI-8200: (see comment below)

N2OM1: column Time

N2Oi2: column Time

SkyLine: column Titles:

uCH4: column Time Stamp

uN2O: column Time Stamp

UGGA: column Time

For the instrument LI-6400, the date is found in one of the first lines in a format containing abbreviations, for example "Thr Aug 6 2020", which would be the date format "mdy". For the instrument LI-8200, the date is found under one of the measurements, next to '"Date":'. For the instrument LI-8250, the date format is in the file name.

The argument proc.data.field is used with the PP-Sytems EGM-5 instrument only. Look up the EGM-5 Operation Manual at page 90 for more details about the different Process Data Fields (proc.data.field).

The arguments active, pivot, flag, PAR.col, Tcham.col, WTD.col, Op.stat.col, inst3, gas3, prec3, dry3, sep, manual and skip are used with the function import.GAIA only. The arguments Tsoil.col, SWC.col, inst1, inst2, gas1, gas2, prec1, prec2, dry1, dry2, are used with both the function import.GAIA and the function import.LI8250. The arguments CH.col and background are used with the function import.GAIA.

The arguments CH4, CO2, sum and range are used with the function import.G2201i only.

The arguments CH.clo.col, sensor1, sensor2, inst and dry are used with the function import.skyline only.

The argument gas is used with the functions import.GasmetPD and import.skyline.

See also

Examples

# Examples on how to use it with all instruments.
# The default time zone "UTC" is used in all cases.

# with the Gasmet instrument DX4015
file.path <- system.file("extdata/DX4015", package = "goFlux")
import2RData(path = file.path, instrument = "DX4015",
             date.format = "ymd", prec = c(1.6, 23, 13, 2, 23, 33))
#> DX4015.TXT saved as DX4015_DX4015_imp.RData in RData folder, in working directory

# with the Gasmet instrument GT5000
file.path <- system.file("extdata/GT5000", package = "goFlux")
import2RData(path = file.path, instrument = "GT5000",
             date.format = "ymd", prec = c(1.6, 23, 13, 2, 23, 33))
#> GT5000.TXT saved as GT5000_GT5000_imp.RData in RData folder, in working directory

# with the PP-Systems EGM-5
# The argument proc.data.field is used with the PP-System EGM-5 instrument only.
file.path <- system.file("extdata/EGM5", package = "goFlux")
import2RData(path = file.path, instrument = "EGM5",
             date.format = "dmy", prec = c(3, 1, 500), proc.data.field = 2)
#> EGM5.TXT saved as EGM5_EGM5_imp.RData in RData folder, in working directory

# with the Picarro instrument G2201i
file.path <- system.file("extdata/G2201i", package = "goFlux")
import2RData(path = file.path, instrument = "G2201i",
             date.format = "ymd", prec = c(0.2, 50, 100))
#> G2201i.dat saved as G2201i_G2201i_imp.RData in RData folder, in working directory

# with the Picarro instrument G2508
file.path <- system.file("extdata/G2508", package = "goFlux")
import2RData(path = file.path, instrument = "G2508",
             date.format = "ymd", prec = c(0.24, 0.3, 5, 0.16, 500))
#> G2508.dat saved as G2508_G2508_imp.RData in RData folder, in working directory

# with the Picarro instrument G4301
file.path <- system.file("extdata/G4301", package = "goFlux")
import2RData(path = file.path, instrument = "G4301",
             date.format = "ymd", prec = c(0.025, 0.1, 10))
#> G4301.dat saved as G4301_G4301_imp.RData in RData folder, in working directory

# with the automated chamber ECOFlux (GAIA2TECH)
# with this instrument, "keep_all" is not a valid argument.
file.path <- system.file("extdata/GAIA", package = "goFlux")
import2RData(path = file.path, instrument = "GAIA", date.format = "ymd",
             pivot = "long", active = TRUE, flag = c(7,11),
             background = FALSE,
             CH.col = "COM5A0",
             SWC.col = "1C08_Soil Moisture",
             Tsoil.col = "1C07_Soil Temperature",
             Tcham.col = "2C07_Chamber Temperature",
             PAR.col = "3C07_Sunlight",
             WTD.col = NULL,
             Op.stat.col = "0C06_OperatingStatus",
             inst1 = "XT2C00_Instrument",
             inst2 = "XT3C00_Instrument",
             gas1 = c("XT2C04_CH4", "XT2C05_CO2", "XT2C06_H2O"),
             gas2 = c("XT3C04_N2O", "XT3C05_H2O"),
             prec1 = c(0.6, 3.5, 45),
             prec2 = c(0.4, 45))
#> GAIA.csv saved as GAIA_GAIA_imp.RData in RData folder, in working directory
#> Warning: The same gas was measured by two instruments. The suffix 'inst1', 'inst2' or 'inst3' was added to each duplicated gas column to avoid confusion. 'inst1' = LICOR78XX#1, 'inst2' = LICOR78XX#2 and 'inst3' = NULL

# with the custom multiplexer from the University of Padova (GasmetPD)
file.path <- system.file("extdata/GasmetPD", package = "goFlux")
import2RData(path = file.path, instrument = "GasmetPD",
             gas = c("CO2_ppm", "CH4_ppm", "N2O_ppm", "NH3_ppm", "H2O_pct"),
             date.format = "ymd", prec = c(1.6, 13, 2, 23, 33))
#> GasmetPD.txt saved as GasmetPD_GasmetPD_imp.RData in RData folder, in working directory

# with Los Gatos Research UGGA (GLA132 series)
file.path <- system.file("extdata/UGGA", package = "goFlux")
import2RData(path = file.path, instrument = "UGGA",
             date.format = "dmy", prec = c(0.2, 1.4, 50))
#> UGGA.txt saved as UGGA_UGGA_imp.RData in RData folder, in working directory

# with Los Gatos Research m-GGA (GLA131 series)
file.path <- system.file("extdata/UGGA", package = "goFlux")
import2RData(path = file.path, instrument = "UGGA",
             date.format = "dmy", prec = c(0.35, 0.9, 200))
#> UGGA.txt saved as UGGA_UGGA_imp.RData in RData folder, in working directory

# with Los Gatos Research N2OM1 (GLA151 series)
file.path <- system.file("extdata/N2OM1", package = "goFlux")
import2RData(path = file.path, instrument = "N2OM1",
             date.format = "dmy", prec = c(0.5, 2, 50))
#> N2OM1.txt saved as N2OM1_N2OM1_imp.RData in RData folder, in working directory

# with Los Gatos Research N2Oi2 (GLA451 series)
file.path <- system.file("extdata/N2Oi2", package = "goFlux")
import2RData(path = file.path, instrument = "N2Oi2",
             date.format = "dmy", prec = c(0.05, 50))
#> N2Oi2.txt saved as N2Oi2_N2Oi2_imp.RData in RData folder, in working directory

# with the LI-COR LI-6400 gas analyzer
file.path <- system.file("extdata/LI6400", package = "goFlux")
import2RData(path = file.path, instrument = "LI-6400",
             date.format = "mdy", prec = c(0.15, 20))
#> LI6400.txt saved as LI6400_LI6400_imp.RData in RData folder, in working directory

# with the LI-COR LI-7810 gas analyzer
file.path <- system.file("extdata/LI7810", package = "goFlux")
import2RData(path = file.path, instrument = "LI-7810",
             date.format = "ymd", prec = c(3.5, 0.6, 45))
#> LI7810.data saved as LI7810_LI7810_imp.RData in RData folder, in working directory

# with the LI-COR LI-7820 gas analyzer
file.path <- system.file("extdata/LI7820", package = "goFlux")
import2RData(path = file.path, instrument = "LI-7820",
             date.format = "ymd", prec = c(0.4, 45))
#> LI7820.data saved as LI7820_LI7820_imp.RData in RData folder, in working directory

# with the LI-COR LI-8100 gas analyzer
file.path <- system.file("extdata/LI8100", package = "goFlux")
import2RData(path = file.path, instrument = "LI-8100",
             date.format = "ymd", prec = c(1, 10))
#> LI8100.81x saved as LI8100_LI8100_imp.RData in RData folder, in working directory

# with the LI-COR smart chamber (LI-8200)
# with this instrument, "keep_all" and "prec" are not valid arguments.
file.path <- system.file("extdata/LI8200", package = "goFlux")
import2RData(path = file.path, instrument = "LI-8200", date.format = "ymd")
#> LI8200.json saved as LI8200_LI8200_imp.RData in RData folder, in working directory

# with the LI-COR Multiplexer (LI-8250)
file.path <- system.file("extdata/LI8250", package = "goFlux")
import2RData(path = file.path, instrument = "LI-8250", date.format = "ymd",
             inst1 = "LI-7810", gas1 = c("CO2_DRY", "CH4_DRY", "H2O"),
             prec1 = c(3.5, 0.6, 45), SWC.col = "SWC_1", Tsoil.col = "TS_1")
#> LI8250.82z saved as LI8250_LI8250_imp.RData in RData folder, in working directory

# with the SkyLine2D system (Earth Bound Scientific)
# with this instrument, "keep_all" is not a valid argument.
file.path <- system.file("extdata/skyline", package = "goFlux")
import2RData(path = file.path, instrument = "skyline", date.format = "ymd",
             deadband = 0,
             shoulder = 20,
             CH.col = "CH ID",
             CH.clo.col = "Chamber closed",
             sensor1 = "Analog Sensor1",
             sensor2 = "Analog Sensor2",
             inst = "Instrument",
             gas = c("CO2", "CH4 Dry", "N2O Dry", "H2O"),
             prec = c(0.24, 0.3, 5, 500))
#> skyline.csv saved as skyline_skyline_imp.RData in RData folder, in working directory

# with the Aeris Technologies MIRA Ultra CH4/C2H6 analyzer
file.path <- system.file("extdata/uCH4", package = "goFlux")
import2RData(path = file.path, instrument = "uCH4",
             date.format = "mdy", prec = c(1, 0.5, 15))
#> uCH4.txt saved as uCH4_uCH4_imp.RData in RData folder, in working directory

# with the Aeris Technologies MIRA Ultra N2O/CO2 analyzer
file.path <- system.file("extdata/uN2O", package = "goFlux")
import2RData(path = file.path, instrument = "uN2O",
             date.format = "mdy", prec = c(0.2, 0.2, 15))
#> uN2O.txt saved as uN2O_uN2O_imp.RData in RData folder, in working directory